requirements.txt
streamlit>=1.31.0
plotly>=5.18.0
pandas>=2.1.0
notion-client>=2.2.0
stripe>=7.0.0
anthropic>=0.39.0
requests>=2.31.0
reportlab>=4.0.0
python-dotenv>=1.0.0
